% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMPProcess.r
\name{CMPProcess}
\alias{CMPProcess}
\title{COM Poisson Process Parameter Estimation}
\usage{
CMPProcess(counts, s = 1)
}
\arguments{
\item{counts}{A vector (like object) of counts}

\item{s}{The time interval corresponding to the levels of the counts vector}
}
\value{
If the time interval (s) is 1, then CMPProcess will return: se=se, H=H, lambda=lambda, nu=nu, Z=Z, ll=loglike, aic=(4-2*loglike).
    
    lambda \deqn{\ln L(\lambda, \nu | x_1,\cdots,x_n)= (\ln \lambda)\sum\limits _{i=1}^n x_i - \nu\sum\limits _{i=1}^n \ln(x_i!) - n\ln Z(\lambda,\nu)}
    nu
    Z
   
    se
    H 
    
    ll
    aic
        
          
   Otherwise, if the time interval (s) is greater than 1, CMPProcess will return lambda=lambda, nu=nu, ll=loglike, aic=(4-2*loglike)
    
    Not sure I'm clear on the form of the ll function of sCMP process. What exactly is the dSCMP function doing?
    
    lambda
    nu
    ll
    aic
}
\description{
\code{CMPPRocess} computes the maximum likelihood estimates of a COM-Poisson process for given count data
}
\examples{

## Standard use
data(floodcount)
CMPProcess(floodcount$Counts)

## Aggregate to s-unit = 3
three.year.bins <- list(floor((floodcount$Year-min(floodcount$Year))/3))
collapsed.floodcount <- aggregate( x = floodcount, by = three.year.bins, FUN=sum)
CMPProcess(collapsed.floodcount$Counts, s= 3)

}
