% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMPProcessWT.r
\name{CMPProcessWT}
\alias{CMPProcessWT}
\title{COM Poisson Process Parameter Estimation}
\usage{
CMPProcessWT(dispersion, t)
}
\arguments{
\item{dispersion}{Ratio between the mean and variance of the process}

\item{t}{A vector (like object) of wait times}
}
\value{
\code{CMPProcessWT} will return a list of three elements: [1] lambda, [2] nu, and [3] convergence
}
\description{
\code{CMPProcessWT} computes the maximum likelihood estimates of a COM-Poisson process for given wait time data
}
\examples{

## Standard use
data(floodcount)
CMPProcessWT(.8 , mean(floodwait$WT))

}
