% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpprocwt.r
\name{cmpprocwt}
\alias{cmpprocwt}
\title{COM Poisson Process Parameter Estimation}
\usage{
cmpprocwt(dispersion, t)
}
\arguments{
\item{dispersion}{Ratio between the mean and variance of the process}

\item{t}{A vector (like object) of wait times}
}
\value{
\code{cmpprocwt} will return a list of three elements: [1] lambda, [2] nu, and [3] convergence
}
\description{
\code{cmpprocwt} computes the maximum likelihood estimates of a COM-Poisson process for given wait time data
}
\examples{

## Standard use
data(floodcount)
cmpprocwt(.8 , mean(floodwait$WT))

}
