% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMPProcess.r
\name{dSCMP}
\alias{dSCMP}
\title{Compute SCMP probabilities.}
\usage{
dSCMP(NuOfVar, Lambda, Nu, count)
}
\arguments{
\item{NuOfVar}{Nu of Var}

\item{Lambda}{Lambda}

\item{Nu}{Nu}

\item{count}{Count}
}
\value{
Will return a probability, given the supplied the parameters
}
\description{
\code{CMPProc} is strictly a helper function that is called by \code{CMPProcess}. This code computes P(X=count) for a sCOM-Poisson(lambda, Nu, NuOfVar) distribution.
}
