% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fetalcount}
\alias{fetalcount}
\title{Fetal lamb movements}
\format{A dataframe with 224 rows and 1 variable
\describe{
 \item{Counts}{Number of movements, note that the variance is 0.693 and the mean is 0.382}
}}
\source{
(Guttorp , 1995)
}
\usage{
fetalcount
}
\description{
Number of movements by a fetal lamb observed by ultrasound and counted in successive 5-second intervals
}
\keyword{datasets}
