% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fetalwait}
\alias{fetalwait}
\title{Fetal lamb movement wait times}
\format{A dataframe with 58 rows and 1 variable
\describe{
 \item{WT}{Number of 5-second intervals between fetal movements}
}}
\usage{
fetalwait
}
\description{
Time between movements by a fetal lamb observed by ultrasound
}
\keyword{datasets}
