% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{floodcount}
\alias{floodcount}
\title{Rio Negro floods}
\format{A dataframe with 101 rows and 2 variables:
\describe{
 \item{Year}{Year of reference}
 \item{Counts}{Count of floods in a given year, note that the variance is .166 and the mean is .208}
}}
\source{
(Brillinger , 1995; Guttorp , 1995)
}
\usage{
floodcount
}
\description{
Years of major floods between 1892-1992 (inclusive) on the Rio Negro River in Brazil
}
\keyword{datasets}
