% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{floodwait}
\alias{floodwait}
\title{Rio Negro flood wait times}
\format{A dataframe with 20 rows and 1 variable:
\describe{
 \item{WT}{Number of years between floods}
}}
\source{
Wait time data were derived from Guttorp, 1995.
}
\usage{
floodwait
}
\description{
Time between major floods between 1892-1992 (inclusive) on the Rio Negro River in Brazil
}
\keyword{datasets}
