% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMPProcess.r
\name{SCMPProc}
\alias{SCMPProc}
\title{COM Poisson Process Parameter Estimation, time interval > 1.}
\usage{
SCMPProc(counts, s, h.out = F)
}
\arguments{
\item{counts}{A vector (like object) of counts}

\item{s}{The time interval corresponding to the levels of the counts vector}

\item{h.out}{A boolean indicating if the information matrix and associated standard errors are returned. This can have a significant impact on processing time and should be left at the default unless needed}
}
\value{
Will return a list of seven elements to \code{CMPProcess}
}
\description{
\code{CMPProc} is strictly a helper function that is called by \code{CMPProcess}
}
