% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMPProcess.r
\name{CMPProcess}
\alias{CMPProcess}
\title{COM Poisson Process Parameter Estimation.}
\usage{
CMPProcess(counts, s = 1, h.out = FALSE)
}
\arguments{
\item{counts}{A vector (-like object) of counts.}

\item{s}{The time interval corresponding to the levels of the counts vector.}

\item{h.out}{A boolean indicating if the information matrix and associated standard errors are returned.
This can have a significant impact on processing time and should be left at the default unless needed}
}
\value{
\code{CMPProcess} will return a list of seven elements: [1] se (Standard errors), [2] H (Information matrix), [3] lambda,
    [4] nu, [5] Z, [6] ll (Log likelihood), and [7] aic.
}
\description{
\code{CMPProcess} computes the maximum likelihood estimates of a COM-Poisson process for given count data.
}
\examples{
## Standard usage
data(floodcount)
CMPProcess(floodcount$Counts)

## Aggregate to s-unit = 3
three.year.bins <- list(floor((floodcount$Year-min(floodcount$Year))/3))
collapsed.floodcount <- aggregate( x = floodcount, by = three.year.bins, FUN=sum)
CMPProcess(collapsed.floodcount$Counts, s= 3)

}
